+++ violet/queue_worker.py
@@ -14,25 +14,33 @@ from .utils import fetch_with_json, fetchrow_with_json
 log = logging.getLogger(__name__)
 
 
-async def _queue_function_wrapper(queue, ctx, args, state=None):
+async def _queue_function_wrapper(queue, ctx, *, fail_mode_state=None):
     """Wrapper for the queue function call.
 
     This wrapper locally manages the declared fail mode of the queue.
-
     """
     try:
-        log.debug("job %s calling with args %r", ctx.job_id, args)
-        await queue.function(ctx, *args)
+        log.debug("job %s calling with args %r", ctx.job_id, ctx.args)
+        await queue.cls.setup(ctx)
+        queue.cls.sched.start_events[str(ctx.job_id)].set()
+        await queue.cls.handle(ctx)
     except Exception as exc:
-        state = state or {}
-        retry = await queue.fail_mode.handle(JobDetails(ctx.job_id), exc, state)
+        fail_mode_state = fail_mode_state or {}
+        retry = await queue.cls.fail_mode.handle(
+            JobDetails(ctx.job_id), exc, fail_mode_state
+        )
         if retry:
-            return await _queue_function_wrapper(queue, ctx, args, state=state)
+            return await _queue_function_wrapper(
+                queue, ctx, fail_mode_state=fail_mode_state
+            )
 
 
-async def release_job(manager, conn, task: asyncio.Task, job_id: str):
+async def release_job(queue, task: asyncio.Task, job_id: str):
     """Release a single job. Fetches the result from it and
     updates the table."""
+    sched = queue.cls.sched
+    conn = sched.db
+
     assert task.done()
 
     new_state = JobState.Completed
@@ -48,13 +56,12 @@ async def release_job(manager, conn, task: asyncio.Task, job_id: str):
     assert new_state in (JobState.Completed, JobState.Error)
 
     log.debug("completed! updating job %s", job_id)
-    queue_name = await conn.fetchval(
-        """
-        UPDATE violet_jobs
+    await conn.execute(
+        f"""
+        UPDATE {queue.cls.name}
         SET state = $1,
             errors = $2
         WHERE job_id = $3
-        RETURNING queue
         """,
         new_state.value,
         new_error,
@@ -62,49 +69,44 @@ async def release_job(manager, conn, task: asyncio.Task, job_id: str):
     )
     log.debug("updated! set job %s => %r", job_id, new_state)
 
-    assert queue_name is not None
-
-    log.debug("check %s in %r", job_id, manager.events)
-    if job_id in manager.events:
+    log.debug("check %s in %r", job_id, sched.events)
+    if job_id in sched.events:
         log.debug("set end event %s", job_id)
-        manager.events[job_id].set()
+        sched.events[job_id].set()
 
     try:
-        manager._poller_sets[queue_name].remove(job_id)
+        sched._poller_sets[queue.cls.name].remove(job_id)
     except KeyError:
         pass
 
 
 async def fetch_jobs(
-    conn,
     queue: Queue,
     state=JobState.NotTaken,
     scheduled_only: bool = False,
     all: bool = False,
 ) -> list:
-    """Fetch a list of jobs based on search parameters."""
-    log.debug("querying state=%r for queue %r", state, queue.name)
+    """Fetch a list of job IDs based on search parameters."""
+    log.debug("querying state=%r for queue %r", state, queue.cls.name)
 
     limit_clause = ""
     if not all:
-        limit_clause = f"LIMIT {queue.poller_rate[0]}"
+        limit_clause = f"LIMIT {queue.cls.poller_rate[0]}"
 
     scheduled_where = (
         "AND (now() at time zone 'utc') >= scheduled_at" if scheduled_only else ""
     )
+
     return await fetch_with_json(
-        conn,
+        queue.cls.sched.db,
         f"""
-        SELECT job_id, name, args
-        FROM violet_jobs
-        WHERE
-            queue = $1
-        AND state = $2
+        SELECT job_id
+        FROM {queue.cls.name}
+        WHERE state = $1
         {scheduled_where}
         ORDER BY inserted_at
         {limit_clause}
         """,
-        queue.name,
         state,
     )
 
@@ -118,26 +120,25 @@ async def fetch_jobs(
 #        await asyncio.sleep(queue.period)
 
 
-async def queue_worker_tick(manager, queue, job_id: Flake):
+async def queue_worker_tick(queue, job_id: Flake):
     """Queue Worker Tick.
 
     This function:
      - Locks the given job (if another worker also takes the job via the
      queue they will fail to lock it properly, it works extremely well for
      multiple workers)
-     - Prepares all eventing (start and stop events)
-     - Runs the queue function
-        (fail modes are handled more locally at the call)
+     - Runs the queue function (fail modes are handled locally, at the call)
      - Fetches the result from the underlying asyncio task and updates the
         database with it.
     """
+    db = queue.cls.sched.db
     row = await fetchrow_with_json(
-        manager.db,
-        """
-        UPDATE violet_jobs
+        db,
+        f"""
+        UPDATE {queue.cls.name}
         SET state = $1, taken_at = (now() at time zone 'utc')
         WHERE job_id = $2 AND state = $3
-        RETURNING args, name
+        RETURNING *
         """,
         JobState.Taken.value,
         str(job_id),
@@ -148,21 +149,17 @@ async def queue_worker_tick(manager, queue, job_id: Flake):
         log.warning("job %r already locked, skipping", job_id)
         return
 
-    ctx = QueueJobContext(manager, queue, job_id, row["name"])
-
-    job_id_str = str(job_id)
-    if not queue.custom_start_event and job_id_str in manager.start_events:
-        manager.start_events[job_id_str].set()
-
-    task = manager.loop.create_task(_queue_function_wrapper(queue, ctx, row["args"]))
+    args = queue.cls.map_persisted_row(row)
+    ctx = QueueJobContext(queue, job_id, row["name"], args)
+    task = queue.cls.sched.loop.create_task(_queue_function_wrapper(queue, ctx))
 
     # TODO add configurable timeout for the tasks?
     await asyncio.wait_for(task, None)
 
-    await release_job(manager, manager.db, task, str(job_id))
+    await release_job(queue, task, str(job_id))
 
 
-async def queue_worker(manager, queue: Queue, worker_id: int):
+async def queue_worker(queue: Queue, worker_id: int):
     """Main queue worker.
 
     This worker keeps waiting for any new job in the backing asyncio queue and
@@ -170,18 +167,18 @@ async def queue_worker(manager, queue: Queue, worker_id: int):
     functionality (e.g locking)
     """
     while True:
-        log.debug("worker %r %d waiting...", queue.name, worker_id)
+        log.debug("worker %r %d waiting...", queue.cls.name, worker_id)
         job_id: Flake = await queue.asyncio_queue.get()
 
-        log.debug("worker %r %d working on %s", queue.name, worker_id, job_id)
-        await queue_worker_tick(manager, queue, job_id)
+        log.debug("worker %r %d working on %s", queue.cls.name, worker_id, job_id)
+        await queue_worker_tick(queue, job_id)
 
 
-def _push_rows_to_queue(manager, queue: Queue, rows: list) -> None:
+def _push_rows_to_queue(queue: Queue, rows: list) -> None:
     """Given a list of rows from fetch_jobs, push them to the queue."""
 
     if rows:
-        log.info("Directly pushing %d jobs to %r", len(rows), queue.name)
+        log.info("Directly pushing %d jobs to %r", len(rows), queue.cls.name)
 
     for row in rows:
         job_id = Flake.from_uuid(row["job_id"])
@@ -189,7 +186,7 @@ def _push_rows_to_queue(manager, queue: Queue, rows: list) -> None:
 
         # prevent pushing the same jobs by putting them in a set and checking
         # if they were already worked on.
-        poller_jobs = manager._poller_sets[queue.name]
+        poller_jobs = queue.cls.sched._poller_sets[queue.cls.name]
         if as_str in poller_jobs:
             continue
 
@@ -198,32 +195,30 @@ def _push_rows_to_queue(manager, queue: Queue, rows: list) -> None:
         poller_jobs.add(as_str)
 
 
-async def run_taken_jobs(manager, queue: Queue):
-    rows = await fetch_jobs(
-        manager.db, queue, state=JobState.Taken, scheduled_only=True, all=True
-    )
+async def run_taken_jobs(queue: Queue):
+    rows = await fetch_jobs(queue, state=JobState.Taken, scheduled_only=True, all=True)
 
     # if any rows were found in the locked state and we are starting the
     # worker, we need to unlock it before pushing can happen.
-    async with manager.db.acquire() as conn:
+    async with queue.cls.sched.db.acquire() as conn:
         for row in rows:
             job_id = Flake.from_uuid(row["job_id"])
             as_str = str(job_id)
 
             log.debug("unlocking job: %s", job_id)
             await conn.execute(
-                """
-                UPDATE violet_jobs
+                f"""
+                UPDATE {queue.cls.name}
                 SET state = 0
                 WHERE job_id = $1
                 """,
                 as_str,
             )
 
-    _push_rows_to_queue(manager, queue, rows)
+    _push_rows_to_queue(queue, rows)
 
 
-async def queue_poller(manager, queue: Queue):
+async def queue_poller(queue: Queue):
     """Queue poller task.
 
     This task queries the database every second and checks all the jobs that
@@ -247,15 +242,13 @@ async def queue_poller(manager, queue: Queue):
     the worker removes the IDs it worked on.
     """
 
-    if queue.start_existing_jobs:
+    if queue.cls.start_existing_jobs:
         log.info("Pushing any locked jobs")
-        await run_taken_jobs(manager, queue)
+        await run_taken_jobs(queue)
         log.info("Pushed any locked jobs")
 
     while True:
-        rows = await fetch_jobs(
-            manager.db, queue, state=JobState.NotTaken, scheduled_only=True
-        )
+        rows = await fetch_jobs(queue, state=JobState.NotTaken, scheduled_only=True)
 
-        _push_rows_to_queue(manager, queue, rows)
-        await asyncio.sleep(queue.poller_rate[1])
+        _push_rows_to_queue(queue, rows)
+        await asyncio.sleep(queue.cls.poller_rate[1])